P2022-600 BoardStatus clean up some code, move Handlers to new location
