+++ src/Patronage.Api/MediatR/BoardStatus/Queries/Handlers/GetByIdBoardStatusHandler.cs
@@ -8,7 +8,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Patronage.Api.MediatR.BoardStatus
+namespace Patronage.Api.MediatR.BoardStatus.Queries.Handlers
 {
     public class GetByIdBoardStatusHandler : IRequestHandler<GetByIdBoardStatusQuery, IEnumerable<BoardStatusDto>>
     {
