+++ src/Patronage.Contracts/ModelDtos/BoardStatusDto.cs
@@ -8,7 +8,6 @@ namespace Patronage.Contracts.ModelDtos
 {
     public class BoardStatusDto
     {
-        //public int Id { get; set; }
         public int BoardId { get; set; }
         public int StatusId { get; set; }
     }
